requirements.txt
ray==1.0.1.post1
ray[tune]
ray[rllib]
ray[debug]
torch==1.7.1
torchvision==0.8.2
tensorflow-gpu==2.3
tensorflow-addons==0.12.1
tensorboardX
gputil
sklearn
matplotlib
gym==0.17.2
cryptography==3.0
luigi==3.0.1
pandas==1.1.1
tabulate==0.8.7
opencv-python==4.4.0.42
pyarrow==1.0.1
smart-open==3.0.0
tqdm==4.51.0
plotly-express==0.4.1
numpy==1.18
git+https://github.com/google/dopamine.git
absl-py 
atari-py 
gin-config 
shap==0.39.0
googledrivedownloader